+++ 14th_homework/maincamp/daily-homework/src/app/boards/[boardId]/page.tsx
@@ -1,115 +1,19 @@
+// 게시글상세 페이지
+
 'use client';
-
-import { gql, useQuery } from '@apollo/client';
-import { useRouter, useParams } from 'next/navigation';
+import CommentWrite from '@/components/boards-detail/comment-write';
 import styles from './styles.module.css';
-import dayjs from 'dayjs';
+import Detail from '@/components/boards-detail/detail';
+import { useParams } from 'next/navigation';
 
-const FETCH_BOARD = gql`
-  query fetchBoard($boardId: ID!) {
-    fetchBoard(boardId: $boardId) {
-      _id
-      writer
-      title
-      contents
-      createdAt
-      updatedAt
-    }
-  }
-`;
-
-const Detail = () => {
-  const myUrl = useParams();
-  const router = useRouter();
-
-  // console.log(myUrl.boardId);
-  const { data } = useQuery(FETCH_BOARD, {
-    variables: {
-      boardId: String(myUrl.boardId),
-    },
-  });
-
-  // console.log(data?.fetchBoard.createdAt);
-
-  const dt = data?.fetchBoard.createdAt;
-  const koreaTime = dayjs(dt).format('YYYY-MM-DD HH:mm:ss');
-  // console.log(koreaTime);
+export default function BoardComponentDetailPage() {
+  const params = useParams();
+  const boardId = params?.boardId as string;
 
   return (
-    <div className={styles['D_layout']}>
-      <div className={styles['D_enroll-row-container']}>
-        <div className={styles['D_title']}>
-          <div className={styles['D_title-text']}>
-            {/* 살어리 살어리랏다 쳥산(靑山)애 살어리랏다멀위랑 ᄃᆞ래랑 먹고 쳥산(靑山)애
-            살어리랏다얄리얄리 얄랑셩 얄라리 얄라 */}
-            {data?.fetchBoard.title}
-          </div>
-          <div className={styles['D_info']}>
-            <div className={styles['D_info-1']}>
-              <div className={styles['D_info-1-profile']}>
-                <svg xmlns="./profile.svg" width="24" height="24" viewBox="0 0 24 24" fill="none">
-                  <rect width="24" height="24" rx="12" fill="#E4E4E4" />
-                  <path
-                    d="M12 11.6924C11.0375 11.6924 10.2136 11.3497 9.52825 10.6644C8.84275 9.97888 8.5 9.15488 8.5 8.19238C8.5 7.22988 8.84275 6.40597 9.52825 5.72063C10.2136 5.03513 11.0375 4.69238 12 4.69238C12.9625 4.69238 13.7864 5.03513 14.4718 5.72063C15.1573 6.40597 15.5 7.22988 15.5 8.19238C15.5 9.15488 15.1573 9.97888 14.4718 10.6644C13.7864 11.3497 12.9625 11.6924 12 11.6924ZM4.5 17.7886V17.0846C4.5 16.595 4.633 16.1415 4.899 15.7241C5.165 15.3068 5.5205 14.986 5.9655 14.7616C6.95383 14.2771 7.95092 13.9137 8.95675 13.6714C9.96258 13.429 10.977 13.3079 12 13.3079C13.023 13.3079 14.0374 13.429 15.0433 13.6714C16.0491 13.9137 17.0462 14.2771 18.0345 14.7616C18.4795 14.986 18.835 15.3068 19.101 15.7241C19.367 16.1415 19.5 16.595 19.5 17.0846V17.7886C19.5 18.2103 19.3523 18.569 19.0568 18.8646C18.7613 19.1601 18.4026 19.3079 17.9808 19.3079H6.01925C5.59742 19.3079 5.23875 19.1601 4.94325 18.8646C4.64775 18.569 4.5 18.2103 4.5 17.7886ZM6 17.8079H18V17.0846C18 16.8821 17.9413 16.6946 17.824 16.5221C17.7067 16.3498 17.5474 16.2091 17.3462 16.1001C16.4846 15.6758 15.6061 15.3543 14.7107 15.1356C13.8153 14.9171 12.9117 14.8079 12 14.8079C11.0883 14.8079 10.1848 14.9171 9.28925 15.1356C8.39392 15.3543 7.51542 15.6758 6.65375 16.1001C6.45258 16.2091 6.29333 16.3498 6.176 16.5221C6.05867 16.6946 6 16.8821 6 17.0846V17.8079ZM12 10.1924C12.55 10.1924 13.0208 9.99655 13.4125 9.60488C13.8042 9.21322 14 8.74238 14 8.19238C14 7.64238 13.8042 7.17155 13.4125 6.77988C13.0208 6.38822 12.55 6.19238 12 6.19238C11.45 6.19238 10.9792 6.38822 10.5875 6.77988C10.1958 7.17155 10 7.64238 10 8.19238C10 8.74238 10.1958 9.21322 10.5875 9.60488C10.9792 9.99655 11.45 10.1924 12 10.1924Z"
-                    fill="#777777"
-                  />
-                </svg>
-                <div className={styles.name}>{data?.fetchBoard.writer}</div>
-              </div>
-              <div className={styles['D_info-1-date']}>{koreaTime}</div>
-            </div>
-            <div className={styles['D_enroll-border']}></div>
-            <div className={styles['D_info-2']}>
-              {/* <div className={styles["D_info-2-icons"]}> */}
-              <svg xmlns="./link.svg" width="24" height="24" viewBox="0 0 24 24" fill="none">
-                <path
-                  d="M7.0385 16.5384C5.78283 16.5384 4.7125 16.0959 3.8275 15.2109C2.9425 14.3261 2.5 13.2559 2.5 12.0004C2.5 10.7449 2.9425 9.67459 3.8275 8.78943C4.7125 7.90409 5.78283 7.46143 7.0385 7.46143H10.0578C10.2702 7.46143 10.4483 7.53334 10.592 7.67718C10.7358 7.82101 10.8077 7.99926 10.8077 8.21192C10.8077 8.42442 10.7358 8.60251 10.592 8.74617C10.4483 8.88967 10.2702 8.96143 10.0578 8.96143H7.03725C6.19825 8.96143 5.48233 9.25792 4.8895 9.85092C4.2965 10.4439 4 11.1603 4 11.9999C4 12.8396 4.2965 13.5559 4.8895 14.1489C5.48233 14.7419 6.19825 15.0384 7.03725 15.0384H10.0578C10.2702 15.0384 10.4483 15.1103 10.592 15.2542C10.7358 15.398 10.8077 15.5762 10.8077 15.7887C10.8077 16.0013 10.7358 16.1794 10.592 16.3229C10.4483 16.4666 10.2702 16.5384 10.0578 16.5384H7.0385ZM9 12.7499C8.7875 12.7499 8.60942 12.678 8.46575 12.5342C8.32192 12.3903 8.25 12.2122 8.25 11.9997C8.25 11.787 8.32192 11.6089 8.46575 11.4654C8.60942 11.3218 8.7875 11.2499 9 11.2499H15C15.2125 11.2499 15.3906 11.3218 15.5343 11.4657C15.6781 11.6095 15.75 11.7877 15.75 12.0002C15.75 12.2128 15.6781 12.3909 15.5343 12.5344C15.3906 12.6781 15.2125 12.7499 15 12.7499H9ZM13.9423 16.5384C13.7298 16.5384 13.5517 16.4665 13.408 16.3227C13.2642 16.1788 13.1923 16.0006 13.1923 15.7879C13.1923 15.5754 13.2642 15.3973 13.408 15.2537C13.5517 15.1102 13.7298 15.0384 13.9423 15.0384H16.9628C17.8018 15.0384 18.5177 14.7419 19.1105 14.1489C19.7035 13.5559 20 12.8396 20 11.9999C20 11.1603 19.7035 10.4439 19.1105 9.85092C18.5177 9.25792 17.8018 8.96143 16.9628 8.96143H13.9423C13.7298 8.96143 13.5517 8.88951 13.408 8.74568C13.2642 8.60184 13.1923 8.42368 13.1923 8.21118C13.1923 7.99851 13.2642 7.82043 13.408 7.67693C13.5517 7.53326 13.7298 7.46143 13.9423 7.46143H16.9615C18.2172 7.46143 19.2875 7.90393 20.1725 8.78893C21.0575 9.67376 21.5 10.7439 21.5 11.9994C21.5 13.2549 21.0575 14.3253 20.1725 15.2104C19.2875 16.0958 18.2172 16.5384 16.9615 16.5384H13.9423Z"
-                  fill="#333333"
-                />
-              </svg>
-              <svg xmlns="./location.svg" width="24" height="24" viewBox="0 0 24 24" fill="none">
-                <path
-                  d="M12.0001 19.5135C13.9565 17.7622 15.4536 16.0823 16.4914 14.474C17.5292 12.8657 18.0481 11.457 18.0481 10.248C18.0481 8.425 17.469 6.92633 16.3106 5.752C15.1523 4.57767 13.7155 3.9905 12.0001 3.9905C10.2848 3.9905 8.84798 4.57767 7.68965 5.752C6.53131 6.92633 5.95215 8.425 5.95215 10.248C5.95215 11.457 6.47107 12.8657 7.5089 14.474C8.54673 16.0823 10.0438 17.7622 12.0001 19.5135ZM12.0001 20.9403C11.8053 20.9403 11.6105 20.9067 11.4156 20.8395C11.2206 20.7722 11.0443 20.668 10.8866 20.527C9.98932 19.7 9.14957 18.8483 8.3674 17.972C7.5854 17.0958 6.90565 16.2199 6.32815 15.3443C5.75048 14.4686 5.2934 13.6007 4.9569 12.7405C4.6204 11.8802 4.45215 11.0493 4.45215 10.248C4.45215 7.94033 5.19856 6.07208 6.6914 4.64325C8.1844 3.21442 9.95398 2.5 12.0001 2.5C14.0463 2.5 15.8159 3.21442 17.3089 4.64325C18.8017 6.07208 19.5481 7.94033 19.5481 10.248C19.5481 11.0493 19.3799 11.8785 19.0434 12.7355C18.7069 13.5927 18.2515 14.4607 17.6771 15.3395C17.1026 16.2183 16.4244 17.0943 15.6424 17.9672C14.8604 18.8404 14.0206 19.6904 13.1231 20.5173C12.9678 20.6583 12.7912 20.764 12.5934 20.8345C12.3957 20.905 12.198 20.9403 12.0001 20.9403ZM12.0019 11.8652C12.4994 11.8652 12.9247 11.6881 13.2779 11.3338C13.6312 10.9794 13.8079 10.5535 13.8079 10.056C13.8079 9.5585 13.6307 9.13308 13.2764 8.77975C12.9221 8.42658 12.4961 8.25 11.9984 8.25C11.5009 8.25 11.0756 8.42717 10.7224 8.7815C10.3691 9.13583 10.1924 9.56183 10.1924 10.0595C10.1924 10.557 10.3696 10.9823 10.7239 11.3355C11.0782 11.6887 11.5042 11.8652 12.0019 11.8652Z"
-                  fill="#333333"
-                />
-              </svg>
-              {/* </div> */}
-            </div>
-          </div>
-        </div>
-        <img src="/image1.jpg" width="400" height="531" />
-        <div className={styles['D_content']}>{data?.fetchBoard.contents}</div>
-        <div className={styles['D_preview']}>
-          <img src="/Frame.png" />
-        </div>
-
-        <div className={styles.Like}>
-          <div className={styles['bad-area']}>
-            <img src="/bad.png" width="24" height="24" />
-            <div>12</div>
-          </div>
-          <div className={styles['good-area']}>
-            <img src="/good.png" width="24" height="24" />
-            <div>24</div>
-          </div>
-        </div>
-
-        <div className={styles['D_button']}>
-          <button className={styles.button}>
-            <img src="/menu.png" width="24" height="24" />
-            <p>목록으로</p>
-          </button>
-          <button
-            className={styles.button}
-            onClick={() => router.push(`/boards/${myUrl.boardId}/edit`)}
-          >
-            <img src="/edit.png" />
-            수정하기
-          </button>
-        </div>
-      </div>
+    <div className={styles.D_layout}>
+      <Detail isEdit={true} ID={boardId} />
+      <CommentWrite />
     </div>
   );
-};
-
-export default Detail;
+}
