+++ 14th_homework/maincamp/daily-homework/src/app/boards/new/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import BoardsWriteAdvanced from '@/commons/components/boards-write';
+import BoardsWriteAdvanced from '@/components/boards-write';
 
 export default function BoardComponentNewPage() {
   return <BoardsWriteAdvanced isEdit={false} />;
