+++ 14th_homework/maincamp/daily-homework/src/app/boards/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import MapBoardsDeletePage from '@/commons/components/boards-list';
+import MapBoardsDeletePage from '@/components/boards-list';
 
 export default function BoardsListPage() {
   return <MapBoardsDeletePage />;
