--- 14th_homework/maincamp/daily-homework/src/commons/components/boards-write/styles.module.css
@@ -1,218 +1,0 @@
-/* * {
-  margin: 0px;
-  padding: 0px;
-} */
-
-.layout {
-  display: flex;
-  width: 1280px;
-  padding: 0 320px;
-  justify-content: center;
-  flex-direction: column;
-  align-items: center;
-  background-color: #fff;
-  /* background-color: lightyellow; */
-}
-
-.enroll-subject {
-  display: flex;
-  width: 1280px;
-  padding: 40px 0px;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: 40px;
-  flex-shrink: 0;
-}
-
-.enroll-subject-text {
-  color: #000;
-
-  /* b_20_28 */
-  font-family: Pretendard;
-  font-size: 20px;
-  font-style: normal;
-  font-weight: 700;
-  line-height: 28px; /* 140% */
-}
-
-.enroll-row-container {
-  display: flex;
-  align-items: flex-start;
-  flex-direction: column;
-  gap: 40px;
-  align-self: stretch;
-  width: 100%;
-}
-
-.enroll-form-title {
-  display: flex;
-  align-items: flex-start;
-  gap: 4px;
-  align-self: stretch;
-}
-
-.enroll-row-section {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: 8px;
-  align-self: stretch;
-  /* width: 100%; */
-}
-
-.enroll-required-indicator {
-  color: var(--red, #f66a6a);
-
-  /* me_16_24 */
-  font-family: 'Pretendard Variable';
-  font-size: 16px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: 24px; /* 150% */
-}
-
-.enroll-input {
-  gap: 8px;
-  width: 100%;
-  padding: 12px 16px;
-  flex-direction: column;
-  align-items: flex-start;
-  flex: 1 0 0;
-
-  border-radius: 8px;
-  border: 1px solid var(--Gray-Gray-200, #d4d3d3);
-  background: var(--Gray-W, #fff);
-  box-sizing: border-box;
-  height: 48px;
-}
-
-.enroll-input[disabled] {
-  gap: 8px;
-  width: 100%;
-  padding: 12px 16px;
-  flex-direction: column;
-  align-items: flex-start;
-  flex: 1 0 0;
-
-  border-radius: 8px;
-  border: 1px solid var(--Gray-Gray-200, #d4d3d3);
-  background-color: lightgray;
-  box-sizing: border-box;
-  height: 48px;
-}
-
-.enroll-border {
-  width: 1280px;
-  height: 1px;
-  background: var(--gray-100, #e4e4e4);
-}
-
-.contents {
-  display: flex;
-  height: 336px;
-  padding: 12px 16px;
-  flex-direction: column;
-  align-items: flex-start;
-  align-self: stretch;
-  border-radius: 8px;
-  border: 1px solid var(--Gray-Gray-200, #d4d3d3);
-  background: var(--Gray-W, #fff);
-}
-
-.enroll-row-flex {
-  display: flex;
-  align-items: flex-start;
-  gap: 40px;
-  align-self: stretch;
-}
-
-.zipcode-input {
-  padding: 12px 16px;
-  box-sizing: border-box;
-  border-radius: 8px;
-  border: 1px solid var(--Gray-Gray-200, #d4d3d3);
-  background: var(--Gray-W, #fff);
-  width: 82px;
-  height: 48px;
-}
-
-.enroll-address-firstrow {
-  display: flex;
-  align-items: center;
-  justify-content: start;
-  align-items: flex-start;
-  gap: 8px;
-  align-self: stretch;
-  height: 48px;
-}
-
-.zipcode-search-button {
-  border-radius: 8px;
-  border: 1px solid var(--gray-B, #000);
-  background: var(--gray-w, #fff);
-  display: flex;
-  height: 48px;
-  padding: 12px 16px;
-  align-items: center;
-  gap: 8px;
-}
-
-.picture-enroll-row {
-  display: flex;
-  align-items: flex-start;
-  gap: 16px;
-}
-
-.enroll-cancel-button {
-  display: flex;
-  height: 48px;
-  padding: 12px 16px;
-  align-items: center;
-  gap: 8px;
-  border-radius: 8px;
-  border: 1px solid var(--gray-B, #000);
-  background: var(--gray-w, #fff);
-}
-
-.enroll-submit-button {
-  border-radius: 8px;
-  background: #2974e5;
-  display: flex;
-  height: 48px;
-  padding: 12px 16px;
-  align-items: center;
-  gap: 8px;
-  border: 0;
-  color: #ffffff;
-}
-
-.enroll-submit-button[disabled] {
-  opacity: 0.5;
-  cursor: not-allowed;
-  background-color: gray;
-}
-
-.enroll-button-container {
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  gap: 16px;
-  align-self: stretch;
-  padding-top: 40px;
-}
-
-.error {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: 8px;
-  flex: 1 0 0;
-  color: red;
-}
-
-.flex-half {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-}
