+++ 14th_homework/maincamp/daily-homework/src/components/boards-detail/comment-list/index.tsx
@@ -1,0 +1,9 @@
+'use client';
+
+import styles from './styles.module.css';
+
+export default function CommentWrite() {
+  return (
+
+  );
+}
