+++ 14th_homework/maincamp/daily-homework/src/components/boards-detail/comment-write/index.tsx
@@ -1,0 +1,97 @@
+'use client';
+
+import styles from './styles.module.css';
+
+export default function CommentWrite() {
+  return (
+    <div className={styles['commentLayout']}>
+      <div className={styles['commentBody']}>
+        <div className={styles['commentTop']}>
+          <svg
+            xmlns="http://www.w3.org/2000/svg"
+            width="24"
+            height="24"
+            viewBox="0 0 24 24"
+            fill="none"
+          >
+            <path
+              d="M6.0385 17.5L4.0365 19.5018C3.752 19.7864 3.42475 19.8509 3.05475 19.6952C2.68492 19.5394 2.5 19.2589 2.5 18.8538V4.30775C2.5 3.80258 2.675 3.375 3.025 3.025C3.375 2.675 3.80258 2.5 4.30775 2.5H19.6923C20.1974 2.5 20.625 2.675 20.975 3.025C21.325 3.375 21.5 3.80258 21.5 4.30775V15.6923C21.5 16.1974 21.325 16.625 20.975 16.975C20.625 17.325 20.1974 17.5 19.6923 17.5H6.0385ZM5.4 16H19.6923C19.7693 16 19.8398 15.9679 19.9038 15.9038C19.9679 15.8398 20 15.7692 20 15.6923V4.30775C20 4.23075 19.9679 4.16025 19.9038 4.09625C19.8398 4.03208 19.7693 4 19.6923 4H4.30775C4.23075 4 4.16025 4.03208 4.09625 4.09625C4.03208 4.16025 4 4.23075 4 4.30775V17.3848L5.4 16Z"
+              fill="#333333"
+            />
+          </svg>{' '}
+          <span className={styles.text}>댓글</span>
+        </div>
+        <div className={styles.rank}>
+          <svg
+            xmlns="http://www.w3.org/2000/svg"
+            width="24"
+            height="24"
+            viewBox="0 0 24 24"
+            fill="none"
+          >
+            <path
+              d="M12 16.688L8.10976 19.034C7.96476 19.1187 7.81951 19.1543 7.67401 19.1407C7.52851 19.1273 7.39684 19.0782 7.27901 18.9938C7.16101 18.9091 7.06993 18.7985 7.00576 18.662C6.94176 18.5255 6.93151 18.3745 6.97501 18.209L8.00776 13.7917L4.57326 10.8187C4.44493 10.7073 4.36318 10.5781 4.32801 10.4313C4.29268 10.2844 4.30193 10.1418 4.35576 10.0033C4.40959 9.86492 4.48718 9.75183 4.58851 9.664C4.68984 9.57617 4.82834 9.51875 5.00401 9.49175L9.53651 9.09575L11.2963 4.9245C11.3603 4.76933 11.4573 4.65458 11.5875 4.58025C11.7177 4.50592 11.8552 4.46875 12 4.46875C12.1448 4.46875 12.2823 4.50592 12.4125 4.58025C12.5427 4.65458 12.6398 4.76933 12.7038 4.9245L14.4635 9.09575L18.996 9.49175C19.1717 9.51875 19.3102 9.57617 19.4115 9.664C19.5128 9.75183 19.5904 9.86492 19.6443 10.0033C19.6981 10.1418 19.7073 10.2844 19.672 10.4313C19.6368 10.5781 19.5551 10.7073 19.4268 10.8187L15.9923 13.7917L17.025 18.209C17.0685 18.3745 17.0583 18.5255 16.9943 18.662C16.9301 18.7985 16.839 18.9091 16.721 18.9938C16.6032 19.0782 16.4715 19.1273 16.326 19.1407C16.1805 19.1543 16.0353 19.1187 15.8903 19.034L12 16.688Z"
+              fill="#C7C7C7"
+            />
+          </svg>
+          <svg
+            xmlns="http://www.w3.org/2000/svg"
+            width="24"
+            height="24"
+            viewBox="0 0 24 24"
+            fill="none"
+          >
+            <path
+              d="M12 16.688L8.10976 19.034C7.96476 19.1187 7.81951 19.1543 7.67401 19.1407C7.52851 19.1273 7.39684 19.0782 7.27901 18.9938C7.16101 18.9091 7.06993 18.7985 7.00576 18.662C6.94176 18.5255 6.93151 18.3745 6.97501 18.209L8.00776 13.7917L4.57326 10.8187C4.44493 10.7073 4.36318 10.5781 4.32801 10.4313C4.29268 10.2844 4.30193 10.1418 4.35576 10.0033C4.40959 9.86492 4.48718 9.75183 4.58851 9.664C4.68984 9.57617 4.82834 9.51875 5.00401 9.49175L9.53651 9.09575L11.2963 4.9245C11.3603 4.76933 11.4573 4.65458 11.5875 4.58025C11.7177 4.50592 11.8552 4.46875 12 4.46875C12.1448 4.46875 12.2823 4.50592 12.4125 4.58025C12.5427 4.65458 12.6398 4.76933 12.7038 4.9245L14.4635 9.09575L18.996 9.49175C19.1717 9.51875 19.3102 9.57617 19.4115 9.664C19.5128 9.75183 19.5904 9.86492 19.6443 10.0033C19.6981 10.1418 19.7073 10.2844 19.672 10.4313C19.6368 10.5781 19.5551 10.7073 19.4268 10.8187L15.9923 13.7917L17.025 18.209C17.0685 18.3745 17.0583 18.5255 16.9943 18.662C16.9301 18.7985 16.839 18.9091 16.721 18.9938C16.6032 19.0782 16.4715 19.1273 16.326 19.1407C16.1805 19.1543 16.0353 19.1187 15.8903 19.034L12 16.688Z"
+              fill="#C7C7C7"
+            />
+          </svg>
+          <svg
+            xmlns="http://www.w3.org/2000/svg"
+            width="24"
+            height="24"
+            viewBox="0 0 24 24"
+            fill="none"
+          >
+            <path
+              d="M12 16.688L8.10976 19.034C7.96476 19.1187 7.81951 19.1543 7.67401 19.1407C7.52851 19.1273 7.39684 19.0782 7.27901 18.9938C7.16101 18.9091 7.06993 18.7985 7.00576 18.662C6.94176 18.5255 6.93151 18.3745 6.97501 18.209L8.00776 13.7917L4.57326 10.8187C4.44493 10.7073 4.36318 10.5781 4.32801 10.4313C4.29268 10.2844 4.30193 10.1418 4.35576 10.0033C4.40959 9.86492 4.48718 9.75183 4.58851 9.664C4.68984 9.57617 4.82834 9.51875 5.00401 9.49175L9.53651 9.09575L11.2963 4.9245C11.3603 4.76933 11.4573 4.65458 11.5875 4.58025C11.7177 4.50592 11.8552 4.46875 12 4.46875C12.1448 4.46875 12.2823 4.50592 12.4125 4.58025C12.5427 4.65458 12.6398 4.76933 12.7038 4.9245L14.4635 9.09575L18.996 9.49175C19.1717 9.51875 19.3102 9.57617 19.4115 9.664C19.5128 9.75183 19.5904 9.86492 19.6443 10.0033C19.6981 10.1418 19.7073 10.2844 19.672 10.4313C19.6368 10.5781 19.5551 10.7073 19.4268 10.8187L15.9923 13.7917L17.025 18.209C17.0685 18.3745 17.0583 18.5255 16.9943 18.662C16.9301 18.7985 16.839 18.9091 16.721 18.9938C16.6032 19.0782 16.4715 19.1273 16.326 19.1407C16.1805 19.1543 16.0353 19.1187 15.8903 19.034L12 16.688Z"
+              fill="#C7C7C7"
+            />
+          </svg>
+          <svg
+            xmlns="http://www.w3.org/2000/svg"
+            width="24"
+            height="24"
+            viewBox="0 0 24 24"
+            fill="none"
+          >
+            <path
+              d="M12 16.688L8.10976 19.034C7.96476 19.1187 7.81951 19.1543 7.67401 19.1407C7.52851 19.1273 7.39684 19.0782 7.27901 18.9938C7.16101 18.9091 7.06993 18.7985 7.00576 18.662C6.94176 18.5255 6.93151 18.3745 6.97501 18.209L8.00776 13.7917L4.57326 10.8187C4.44493 10.7073 4.36318 10.5781 4.32801 10.4313C4.29268 10.2844 4.30193 10.1418 4.35576 10.0033C4.40959 9.86492 4.48718 9.75183 4.58851 9.664C4.68984 9.57617 4.82834 9.51875 5.00401 9.49175L9.53651 9.09575L11.2963 4.9245C11.3603 4.76933 11.4573 4.65458 11.5875 4.58025C11.7177 4.50592 11.8552 4.46875 12 4.46875C12.1448 4.46875 12.2823 4.50592 12.4125 4.58025C12.5427 4.65458 12.6398 4.76933 12.7038 4.9245L14.4635 9.09575L18.996 9.49175C19.1717 9.51875 19.3102 9.57617 19.4115 9.664C19.5128 9.75183 19.5904 9.86492 19.6443 10.0033C19.6981 10.1418 19.7073 10.2844 19.672 10.4313C19.6368 10.5781 19.5551 10.7073 19.4268 10.8187L15.9923 13.7917L17.025 18.209C17.0685 18.3745 17.0583 18.5255 16.9943 18.662C16.9301 18.7985 16.839 18.9091 16.721 18.9938C16.6032 19.0782 16.4715 19.1273 16.326 19.1407C16.1805 19.1543 16.0353 19.1187 15.8903 19.034L12 16.688Z"
+              fill="#C7C7C7"
+            />
+          </svg>
+          <svg
+            xmlns="http://www.w3.org/2000/svg"
+            width="24"
+            height="24"
+            viewBox="0 0 24 24"
+            fill="none"
+          >
+            <path
+              d="M12 16.688L8.10976 19.034C7.96476 19.1187 7.81951 19.1543 7.67401 19.1407C7.52851 19.1273 7.39684 19.0782 7.27901 18.9938C7.16101 18.9091 7.06993 18.7985 7.00576 18.662C6.94176 18.5255 6.93151 18.3745 6.97501 18.209L8.00776 13.7917L4.57326 10.8187C4.44493 10.7073 4.36318 10.5781 4.32801 10.4313C4.29268 10.2844 4.30193 10.1418 4.35576 10.0033C4.40959 9.86492 4.48718 9.75183 4.58851 9.664C4.68984 9.57617 4.82834 9.51875 5.00401 9.49175L9.53651 9.09575L11.2963 4.9245C11.3603 4.76933 11.4573 4.65458 11.5875 4.58025C11.7177 4.50592 11.8552 4.46875 12 4.46875C12.1448 4.46875 12.2823 4.50592 12.4125 4.58025C12.5427 4.65458 12.6398 4.76933 12.7038 4.9245L14.4635 9.09575L18.996 9.49175C19.1717 9.51875 19.3102 9.57617 19.4115 9.664C19.5128 9.75183 19.5904 9.86492 19.6443 10.0033C19.6981 10.1418 19.7073 10.2844 19.672 10.4313C19.6368 10.5781 19.5551 10.7073 19.4268 10.8187L15.9923 13.7917L17.025 18.209C17.0685 18.3745 17.0583 18.5255 16.9943 18.662C16.9301 18.7985 16.839 18.9091 16.721 18.9938C16.6032 19.0782 16.4715 19.1273 16.326 19.1407C16.1805 19.1543 16.0353 19.1187 15.8903 19.034L12 16.688Z"
+              fill="#C7C7C7"
+            />
+          </svg>
+        </div>
+        <div className={styles.commentMiddle}>
+          <div className={styles.commentInfo}>
+            <div className={styles['enroll-row']}>
+              <div className={styles['enroll-input']}></div>
+              <div className={styles['enroll-input']}></div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+}
