+++ 14th_homework/maincamp/daily-homework/src/components/boards-detail/comment-write/styles.module.css
@@ -1,0 +1,95 @@
+.commentLayout {
+  display: flex;
+  width: 1280px;
+  height: 460px;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 40px;
+  border: solid 1px blue;
+  /* background-color: lightblue; */
+}
+
+.commentBody {
+  display: flex;
+  width: 1280px;
+  height: 400px;
+  flex-direction: column;
+  gap: 24px;
+  align-items: stretch;
+  border: solid 1px red;
+
+  /* background-color: yellow; */
+}
+
+.commentTop {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  width: 60px;
+  height: 24px;
+  border: solid 1px pink;
+}
+
+.text {
+  /* sb_16_24 */
+  font-family: 'Pretendard Variable';
+  font-size: 16px;
+  font-style: normal;
+  font-weight: 600;
+  line-height: 24px; /* 150% */
+}
+
+.rank {
+  display: flex;
+  height: 24px;
+  align-items: center;
+  gap: 8px;
+  align-self: stretch;
+}
+
+.commentMiddle {
+  display: flex;
+  height: 304px;
+  flex-direction: column;
+  justify-content: flex-start;
+  align-items: flex-end;
+  gap: 16px;
+  align-self: stretch;
+  border: solid 1px red;
+}
+
+.commentInfo {
+  height: 240px;
+  display: flex;
+  flex-direction: column;
+  /* justify-content: center; */
+  align-items: flex-start;
+  gap: 16px;
+  align-self: stretch;
+  border: solid 1px skyblue;
+}
+
+.enroll-row {
+  display: flex;
+  justify-content: flex-start;
+  width: 640px;
+  height: 80px;
+  /* align-items: flex-start; */
+  gap: 16px;
+  border: solid 1px green;
+}
+
+.enroll-input {
+  width: 312px;
+  height: 80px;
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 8px;
+  flex: 1 0 0;
+
+  border: solid 1px blue;
+}
+
+.commentButton {
+}
