+++ 14th_homework/maincamp/daily-homework/src/components/boards-detail/detail/queries.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import { gql } from '@apollo/client';
 
 export const FETCH_BOARD = gql`
