+++ 14th_homework/maincamp/daily-homework/src/components/boards-detail/detail/types.ts
