+++ 14th_homework/maincamp/daily-homework/src/components/boards-list/queries.ts
