+++ 14th_homework/maincamp/daily-homework/src/components/boards-write/index.tsx
@@ -1,6 +1,5 @@
 'use client';
 import styles from './styles.module.css';
-// import styles from '.';
 import useBoardsWriteAdvanced from './hook';
 import { BoardVariables } from './types';
