+++ 14th_homework/maincamp/daily-homework/src/components/boards-write/types.ts
